+++ tests/integration/test_cmd_script.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
 import os
+import platform
 import shutil
 import subprocess
 import time
 import unittest
 
-import gamspy.utils as utils
 from gamspy import Container
 
 try:
@@ -16,11 +16,19 @@ try:
 except Exception:
     pass
 
+user_dir = os.path.expanduser("~")
+if platform.system() == "Linux":
+    DEFAULT_DIR = os.path.join(user_dir, ".local", "share", "GAMSPy")
+elif platform.system() == "Darwin":
+    DEFAULT_DIR = os.path.join(
+        user_dir, "Library", "Application Support", "GAMSPy"
+    )
+elif platform.system() == "Windows":
+    DEFAULT_DIR = os.path.join(user_dir, "Documents", "GAMSPy")
+
 
 class CmdSuite(unittest.TestCase):
     def test_install_license(self):
-        gamspy_base_directory = utils._get_gamspy_base_directory()
-
         # Test network license
         _ = subprocess.run(
             [
@@ -32,11 +40,9 @@ class CmdSuite(unittest.TestCase):
             check=True,
         )
 
-        user_license_path = os.path.join(
-            gamspy_base_directory, "user_license.txt"
-        )
+        gamspy_license_path = os.path.join(DEFAULT_DIR, "gamspy_license.txt")
 
-        self.assertTrue(os.path.exists(user_license_path))
+        self.assertTrue(os.path.exists(gamspy_license_path))
 
         m = Container()
         self.assertTrue(m._network_license)
@@ -62,8 +68,8 @@ class CmdSuite(unittest.TestCase):
             )
 
         # Test installing a license from a file path.
-        tmp_license_path = os.path.join("tmp", "user_license.txt")
-        shutil.copy(user_license_path, tmp_license_path)
+        tmp_license_path = os.path.join("tmp", "gamspy_license.txt")
+        shutil.copy(gamspy_license_path, tmp_license_path)
 
         _ = subprocess.run(
             [
@@ -84,20 +90,16 @@ class CmdSuite(unittest.TestCase):
             check=True,
         )
 
-        self.assertTrue(os.path.exists(user_license_path))
+        self.assertTrue(os.path.exists(gamspy_license_path))
 
     def test_uninstall_license(self):
-        gamspy_base_directory = utils._get_gamspy_base_directory()
-
         _ = subprocess.run(
             ["gamspy", "uninstall", "license"],
             check=True,
         )
 
         self.assertFalse(
-            os.path.exists(
-                os.path.join(gamspy_base_directory, "user_license.txt")
-            )
+            os.path.exists(os.path.join(DEFAULT_DIR, "gamspy_license.txt"))
         )
 
     def test_install_solver(self):
@@ -204,26 +206,10 @@ class CmdSuite(unittest.TestCase):
             check=True,
         )
 
-        # Copy license to the default path
-        gamspy_base_directory = utils._get_gamspy_base_directory()
-        license_path = os.path.join(gamspy_base_directory, "user_license.txt")
-        default_path = os.path.join(
-            os.path.expanduser("~"), "gamspy_license.txt"
-        )
-        shutil.copy(license_path, default_path)
-
-        # Delete the license from gamspy_base directory
-        _ = subprocess.run(
-            [
-                "gamspy",
-                "uninstall",
-                "license",
-            ],
-            check=True,
-        )
-
         m = Container()
-        self.assertEqual(m._license_path, default_path)
+        self.assertEqual(
+            m._license_path, os.path.join(DEFAULT_DIR, "gamspy_license.txt")
+        )
 
     @classmethod
     def tearDownClass(cls):