+++ CHANGELOG.md
@@ -5,7 +5,7 @@ GAMSPy 1.0.1
 - General
   - Fix frozen solve with non-scalar symbols.
   - Fix the definition update problem while redefining an equation with definition argument.
-  - Allow users to put their licenses to a default path to avoid reinstallation on gamspy_base upgrades.
+  - Introduce default directories to keep license information on upgrade.
 - Testing
   - Add test for the frozen solve with non-scalar symbols.
   - Add a test to verify the behaviour of equation redefinition with definition argument.